requirements.txt
fastapi
uvicorn
torch
torchvision
ffmpeg-python
numpy
requests
opencv-python
tqdm
matplotlib
pydantic
PyYAML
prometheus-client
pytest
ruff
black
mypy
